requirements.txt
beautifulsoup4==4.13.1
numpy==2.2.2
pandas==2.2.3
plotly==5.24.1
prophet==1.1.5
pyecharts==2.0.8
pywebview==5.4
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
streamlit==1.41.1
streamlit_echarts==0.4.0
streamlit_searchbox==0.1.20
yfinance==0.2.52
